DOC/src/backend/requirements.txt
Django==5.1.3
djangorestframework==3.15.2
django-cors-headers==4.6.0
python-decouple==3.8
Pillow==11.0.0
psycopg2-binary==2.9.10
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
celery==5.4.0
redis==5.2.0
django-filter==24.3
